requirements.txt
tqdm
bcrypt
streamlit==1.30.0
opencv-python-headless==4.8.1.78
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.2.2
tqdm>=4.65.0
